Support for experimental SugarCube versions
First, get a branch of the SugarCube repository (for example `git clone
https://github.com/tmedwards/sugarcube-2`). Ensure the folder
`sugarcube-2` is in the same parent folder as this project. If you wish to
use a different project layout, then you can use your own virtual story format
by making a symbol link in the `devTools/storyFormats/` folder.

Compile with the flag `-f sugarcube-2-experimental`.

This change also includes support for the latest version of SugarCube
(2.33.2) which had a small breaking change.