Added Robbing Dojo Cash Register
Forgot to add these because they were in a different file. In the future I may make this more elaborate, since I'm sure Aki would beat the shit out of anyone who robbed the joint if they got caught, but for now it's good enough to not get error messages.