Started Adding Farmhouse Code
Added new events allowing Alex to ask the player to stay the night at the farmhouse if they are dating. (Should be also based on Alex's love, but that code isn't working for some reason.)

Also added the beginnings of code allowing player to sleep at the Farmhouse! Still need to do a ton of work on this, but the start is there. (Need to add a scene where Alex gives you the key, add how said key affects things like Alex being locked in the house, make a button so you can get in when not directly invited, etc.