Continuing to Add Farmhouse Code
Added a metric shitton of stuff. Fixed broken links and if clauses. Implemented Alex giving you the farmhouse key, which allows you to enter the farmhouse at any time going forward. Also made it so you can use the shower in the farmhouse shed, but only after getting the key because it represents having permission to use the shower. (Later I will make this compatible with the Alex showering scene.)