Added Road Pet Sex Scene + Misc
New scene for having sex with the road pet/getting raped by the road pet. Also added transform points for cow when you milk cattle with your mouth. 

Also, added deviancy to having sex with the fox in the meadow. Forgot about deviancy when I added this. 

Think I reverted the change so chance for cow to molest you when you sleep in the barn is no longer divided by two. Honestly, I think sleeping in the barn should be a bit more dangerous, and it's not like it triggers more often than any other farm danger.