Allow "Rest near the stables (0:15)" after stage 7 when there are not at least 2 farm tasks available.
There are a number of ways to run out of tasks (e.g. around 19:30 while Alex is doing the last evening task), so this will give the player a nice way of skipping some time.

I used "at least 2" because it is reasonable to have one task you just do not want to do.