Color the item and amount market buyers pick up
I found that hard to find visually when tending a stall the whole day. After a while you don't read the text anymore but just scan for the item.

Also added more flavour words for the amounts. People still use those in these kinds of settings. (I managed to resist the urge to go full "two gross, 3 score, a dozen and 3"...)