Fix for bug Narcolepsy recently found with Whitney genital parasite event not incrementing herm rep
Narcolepsy recently posted a screenshot of their hermaphrodite character getting the event where Whitney puts a parasite on their genitals, and then mentioned that it doesn't increment herm rep.

I looked into it and discovered that beyond that, it also doesn't seem to set Whitney as knowing the player's gender. So I added functionality to set Whitney as knowing the player's gender and also to increment any relevant reputations (using the schoolrep_naked function instead of manually).

I didn't do anything more involved that may or may not be called for, like having Whitney react to the PC's genitals if they're unexpected, or having a crowd form in response; this was intended purely to be a "bugfix." If you want me to flesh it out a tiny bit more, though, I could take a crack at it.