Bugfix for the Whitney toilets initial event not accounting for herms
Another Narcolepsy bug, this time related to the game not accounting for herms in a particular piece of event text. In the initial Whitney toilets event, when Whitney shoves you into the bathroom, it doesn't account for herms and drops them into the <<else>> case, which is the boys' bathroom, when it should actually be the same bathroom that Whitney uses. I went ahead and fixed this, adding an <<elseif>> for if the player is a herm and using the <<nnpc_gendery "Whitney">> widget to fetch Whitney's pronoun gender. I don't think this presently needs any further updating *unless* you think Whitney should force Obligate Crossdressers to use the bathroom of their birth gender instead of the gender they prefer to pass as, which would make sense.