Base for location overhaul
This is the base for the upcoming changes on how $location, $outside, $bus and tanning values are handled.

The new code currently only sets up the location system, it is not yet used. This requires another change that touches almost every passage (but can be handled by search&replace for almost everything).